backend/data_cleaning/requirements.txt
# Core dependencies
python-dotenv>=0.19.0
celery>=5.2.0
redis>=4.0.0

# String matching and fuzzy search
python-Levenshtein>=0.12.0
fuzzywuzzy>=0.18.0
python-rapidfuzz>=2.0.0

# Database
supabase>=0.7.0
pandas>=1.3.0

# Utilities
tqdm>=4.62.0
tabulate>=0.8.9 